refactor(utils): move protocols to usecase layer
