+++ packages/sp-proxy/src/interface-adapters/utils/formatters/GenerateMetadataFormatter.ts
@@ -1,5 +1,5 @@
-import { IKeyCertFormatter } from '@sp-proxy/interface-adapters/protocols/IKeySetFormatter'
-import { BaseKeyCertFormatter } from '@sp-proxy/interface-adapters/utils/formatters/BaseKeyCertFormatter'
+import { BaseKeyCertFormatter } from '@sp-proxy/use-cases/protocols/BaseKeyCertFormatter'
+import { IKeyCertFormatter } from '@sp-proxy/use-cases/protocols/IKeySetFormatter'
 
 export class GenerateMetadataFormatter
   extends BaseKeyCertFormatter
