+++ packages/sp-proxy/src/use-cases/protocols/BaseKeyCertFormatter.test.ts
@@ -1,4 +1,4 @@
-import { BaseKeyCertFormatter } from '@sp-proxy/interface-adapters/utils/formatters/BaseKeyCertFormatter'
+import { BaseKeyCertFormatter } from '@sp-proxy/use-cases/protocols/BaseKeyCertFormatter'
 import { readFileSync } from 'fs'
 
 const loadedCert = readFileSync(
