+++ packages/sp-proxy/src/use-cases/protocols/IKeyCertLoader.ts
